fix(Tensor): cast rank to int32_t for negative index handling
fix(Conv2DOp): add missing semicolon and improve bias registration logic
feat(Conv2DOp): add runtime assertions for kernel, stride, padding, and dilation sizes

fix(LayerNorm2DOp): improve parameter registration with shared init region

fix(deepseek_ocr): move batch_size declaration after pixel_values assignment
fix(deepseek_ocr): correct height_crop_num indexing in spatial crop usage

feat(Tensor): implement clone method for tensor copying

+++ mllm/core/Tensor.cpp
@@ -347,7 +347,7 @@ bool Tensor::isContiguousN(int n) const { return impl()->isContiguousN(n); }
 
 int32_t Tensor::size(int32_t id) const {
   auto nid = id;
-  if (id < 0) { nid = rank() + id; }
+  if (id < 0) { nid = static_cast<int32_t>(rank()) + id; }
   return shape()[nid];
 }
 
@@ -436,6 +436,7 @@ Tensor Tensor::flatten(int32_t dim) {
 
   return view(new_shape);
 }
+
 Tensor Tensor::clone() { return Context::instance().buildOpAndSubmitTask(OpTypes::kClone, aops::CloneOpOptions{}, {*this})[0]; }
 
 void Tensor::copy2(const Tensor& src) {
